Add `stim::DemInstruction` repeat block helper methods - Extract `dem_instruction.{h,cc}` files - Fix test util `run_captured_stim_main` not having a way to specify `--in` for an empty file anymore
