+++ file_lists/source_files_no_main
@@ -29,6 +29,7 @@ src/stim/cmd/command_m2d.cc
 src/stim/cmd/command_repl.cc
 src/stim/cmd/command_sample.cc
 src/stim/cmd/command_sample_dem.cc
+src/stim/dem/dem_instruction.cc
 src/stim/dem/detector_error_model.cc
 src/stim/diagram/ascii_diagram.cc
 src/stim/diagram/base64.cc
