+++ src/stim/dem/detector_error_model.h
@@ -24,51 +24,9 @@
 
 #include "stim/circuit/circuit.h"
 #include "stim/mem/monotonic_buffer.h"
+#include "stim/dem/dem_instruction.h"
 
 namespace stim {
-
-enum DemInstructionType : uint8_t {
-    DEM_ERROR,
-    DEM_SHIFT_DETECTORS,
-    DEM_DETECTOR,
-    DEM_LOGICAL_OBSERVABLE,
-    DEM_REPEAT_BLOCK,
-};
-
-struct DemTarget {
-    uint64_t data;
-
-    static DemTarget observable_id(uint64_t id);
-    static DemTarget relative_detector_id(uint64_t id);
-    static constexpr DemTarget separator() {
-        return {UINT64_MAX};
-    }
-    uint64_t raw_id() const;
-    uint64_t val() const;
-    bool is_observable_id() const;
-    bool is_separator() const;
-    bool is_relative_detector_id() const;
-    void shift_if_detector_id(int64_t offset);
-
-    bool operator==(const DemTarget &other) const;
-    bool operator!=(const DemTarget &other) const;
-    bool operator<(const DemTarget &other) const;
-    std::string str() const;
-};
-
-struct DemInstruction {
-    SpanRef<const double> arg_data;
-    SpanRef<const DemTarget> target_data;
-    DemInstructionType type;
-
-    bool operator<(const DemInstruction &other) const;
-    bool operator==(const DemInstruction &other) const;
-    bool operator!=(const DemInstruction &other) const;
-    bool approx_equals(const DemInstruction &other, double atol) const;
-    std::string str() const;
-
-    void validate() const;
-};
 
 struct DetectorErrorModel {
     MonotonicBuffer<double> arg_buf;
@@ -146,8 +104,8 @@ struct DetectorErrorModel {
                     callback(DemInstruction{op.arg_data, translate_buf, op.type});
                     break;
                 case DEM_REPEAT_BLOCK: {
-                    const auto &block = blocks[op.target_data[1].data];
-                    auto reps = op.target_data[0].data;
+                    const auto &block = op.repeat_block_body(*this);
+                    auto reps = op.repeat_block_rep_count();
                     for (uint64_t k = 0; k < reps; k++) {
                         block.iter_flatten_error_instructions_helper(callback, detector_shift);
                     }
