+++ src/stim/dem/detector_error_model.pybind.cc
@@ -351,7 +351,7 @@ void stim_pybind::pybind_detector_error_model_methods(
             auto &op = self.instructions[index];
             if (op.type == DEM_REPEAT_BLOCK) {
                 return pybind11::cast(
-                    ExposedDemRepeatBlock{op.target_data[0].data, self.blocks[op.target_data[1].data]});
+                    ExposedDemRepeatBlock{op.repeat_block_rep_count(), op.repeat_block_body(self)});
             }
             ExposedDemInstruction result;
             result.targets.insert(result.targets.begin(), op.target_data.begin(), op.target_data.end());
